rename filename and update readme
