+++ configs/datasets/MMLUArabic/MMLUArabic_gen.py
@@ -1,4 +1,4 @@
 from mmengine.config import read_base
 
 with read_base():
-    from .MMLUArabic_gen_f5fb38 import MMLUArabic_datasets  # noqa: F401, F403
+    from .MMLUArabic_gen_326684 import MMLUArabic_datasets  # noqa: F401, F403
