+++ configs/datasets/MMLUArabic/MMLUArabic_gen_326684.py
