+++ configs/datasets/MMLUArabic/MMLUArabic_zero_shot_gen_3523e0.py
